blender_scoring_service/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.115.6
    # via -r .\requirements.in
filelock==3.16.1
    # via torch
fsspec==2024.12.0
    # via torch
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.5
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
pydantic==2.10.3
    # via
    #   -r .\requirements.in
    #   fastapi
pydantic-core==2.27.1
    # via pydantic
python-dotenv==1.0.1
    # via -r .\requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.41.3
    # via fastapi
sympy==1.13.1
    # via torch
torch==2.5.1
    # via -r .\requirements.in
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   torch
uvicorn==0.34.0
    # via -r .\requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
